add /data volume in the example, use actual path in the example
If a user is just following these commands, the `-v /data` needs to be added so it works
The /var/www/project is what is used in the docker compose snippet.

One thing to note is that with these examples and the current code the `/unison` volume
gets mounted along with `/data`.